requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
APScheduler==3.9.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
Cython==3.0a4
frozenlist==1.3.0
google-api-core==2.8.2
google-auth==2.9.0
google-cloud-speech==2.14.1
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
multidict==6.0.2
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytesseract==0.3.9
python-telegram-bot==20.0a2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
yarl==1.7.2
